# rigweda/backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rigweda API
  version: 0.8.0
  description: Independently implemented HR and payroll platform API.
servers:
  - url: /api/v1
paths:
...
  /access/users:
    get:
      tags: [Access control]
      summary: List users in the authenticated organization
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive, all] } }
      responses:
        "200": { description: Paginated organization-user collection }
        "403": { description: Users-manage permission required }
    post:
      tags: [Access control]
      summary: Create or attach a user and assign an organization role
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: User membership created }
        "409": { description: User is already an organization member }
  /access/users/{userId}:
    patch:
      tags: [Access control]
      summary: Update a user membership or role
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: User membership updated }
        "404": { description: User membership not found }
  /access/roles:
    get:
      tags: [Access control]
      summary: List organization roles and assigned permissions
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Organization roles }
    post:
      tags: [Access control]
      summary: Create a custom role
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Role created }
        "409": { description: Role name conflict }
  /access/roles/{roleId}:
    patch:
      tags: [Access control]
      summary: Update a role and its permissions
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: roleId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Role updated }
        "422": { description: Protected role cannot be reduced }
  /access/permissions:
    get:
      tags: [Access control]
      summary: List the permission catalog
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Permission catalog }
...
  /attendance/metadata:
    get:
      tags: [Attendance]
      summary: Retrieve employees and active shifts for attendance forms
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Attendance form metadata }
  /attendance/me/today:
    get:
      tags: [Attendance]
      summary: Retrieve the current employee's shift and attendance state
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current organization date, assigned shift, and attendance record }
        "404": { description: Login has no linked employee profile }
  /attendance/me/punch:
    post:
      tags: [Attendance]
      summary: Record an immutable clock or break punch
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [clock_in, break_start, break_end, clock_out] }
                latitude: { type: [number, "null"], minimum: -90, maximum: 90 }
                longitude: { type: [number, "null"], minimum: -180, maximum: 180 }
                deviceInfo: { type: [string, "null"], maxLength: 300 }
      responses:
        "200": { description: Punch recorded and attendance totals recalculated }
        "409": { description: Punch action is invalid for the current state }
        "422": { description: Policy or geolocation requirement failed }
  /attendance/me/regularizations:
    get:
      tags: [Attendance]
      summary: List the current employee's correction requests
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Paginated self-service correction requests }
    post:
      tags: [Attendance]
      summary: Request an attendance correction
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Correction request submitted }
        "409": { description: A pending request already exists for this attendance date }
        "422": { description: Correction policy or time window validation failed }
  /attendance/me/regularizations/{requestId}/cancel:
    post:
      tags: [Attendance]
      summary: Cancel a pending self-service correction request
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: requestId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Pending request cancelled }
        "409": { description: Optimistic version conflict or request no longer pending }
  /attendance/daily/summary:
    get:
      tags: [Attendance]
      summary: Retrieve attendance totals and exception counts for one date
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: Daily present, absent, late, weekly-off, and open-punch totals }
  /attendance/daily:
    get:
      tags: [Attendance]
      summary: Search the employee attendance register for one date
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [not_marked, present, absent, half_day, on_leave, holiday, weekly_off, missing_punch, late, overtime, open_punch, all] } }
      responses:
        "200": { description: Paginated daily attendance register }
  /attendance/records:
    get:
      tags: [Attendance]
      summary: Search historical attendance records
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: dateFrom, in: query, required: true, schema: { type: string, format: date } }
        - { name: dateTo, in: query, required: true, schema: { type: string, format: date } }
        - { name: employeeId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200": { description: Paginated historical attendance records }
    put:
      tags: [Attendance]
      summary: Create or correct an attendance record with optimistic locking
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Attendance record corrected and audit history appended }
        "409": { description: Optimistic version conflict }
  /attendance/records/export:
    get:
      tags: [Attendance]
      summary: Export filtered attendance records as spreadsheet-safe CSV
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: dateFrom, in: query, required: true, schema: { type: string, format: date } }
        - { name: dateTo, in: query, required: true, schema: { type: string, format: date } }
      responses:
        "200":
          description: UTF-8 attendance CSV export
          content: { text/csv: { schema: { type: string } } }
  /attendance/records/{recordId}:
    get:
      tags: [Attendance]
      summary: Retrieve one attendance record with immutable punches
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: recordId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Attendance record and punch timeline }
        "404": { description: Attendance record not found }
  /attendance/records/{recordId}/history:
    get:
      tags: [Attendance]
      summary: Retrieve the manual-correction audit history
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: recordId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Immutable attendance change history }
  /attendance/policy:
    get:
      tags: [Attendance]
      summary: Retrieve the organization attendance policy
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Attendance thresholds and self-service policy }
    patch:
      tags: [Attendance]
      summary: Update attendance calculations, weekends, and punch policy
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Attendance policy updated }
        "409": { description: Optimistic version conflict }
  /attendance/shifts:
    get:
      tags: [Attendance]
      summary: Search shift templates
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Paginated shift templates with assignment counts }
    post:
      tags: [Attendance]
      summary: Create a reusable shift template
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Shift template created }
        "409": { description: Shift code conflict }
  /attendance/shifts/{shiftId}:
    parameters:
      - { name: shiftId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Attendance]
      summary: Retrieve a shift template
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Shift template details }
    patch:
      tags: [Attendance]
      summary: Update, activate, deactivate, or make a shift the default
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Shift template updated }
        "409": { description: Optimistic version conflict }
  /attendance/assignments:
    get:
      tags: [Attendance]
      summary: Search date-effective employee shift assignments
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Paginated shift assignments }
    post:
      tags: [Attendance]
      summary: Assign an employee to a shift for an effective period
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Shift assignment created }
        "422": { description: Assignment dates overlap another employee shift }
  /attendance/assignments/{assignmentId}:
    patch:
      tags: [Attendance]
      summary: Change or end an employee shift assignment
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: assignmentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Shift assignment updated }
        "409": { description: Optimistic version conflict }
        "422": { description: Assignment dates overlap another employee shift }
  /attendance/regularizations:
    get:
      tags: [Attendance]
      summary: Search attendance correction requests for approval
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Paginated organization correction requests }
  /attendance/regularizations/{requestId}/review:
    post:
      tags: [Attendance]
      summary: Approve or reject an attendance correction request
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: requestId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Request reviewed; approved corrections applied and audited }
        "409": { description: Optimistic version conflict or request no longer pending }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...