# rigweda/backend/src/modules/access/access.routes.js
const express=require("express");const {asyncHandler}=require("../../middleware/async-handler");const {validate}=require("../../middleware/validate");const schemas=require("./access.validation");
const createAccessRouter=(service)=>{const router=express.Router();router.get("/users",validate(schemas.listUsersSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.listUsers(req.auth.organizationId,req.validated.query)})));router.post("/users",validate(schemas.createUserSchema),asyncHandler(async(req,res)=>res.status(201).json({success:true,data:await service.createUser(req.auth.organizationId,req.validated.body)})));router.patch("/users/:userId",validate(schemas.updateUserSchema),asyncHandler(async(req,res)=>{await service.updateUser(req.auth.organizationId,req.validated.params.userId,req.validated.body);res.status(204).end();}));router.get("/roles",asyncHandler(async(req,res)=>res.json({success:true,data:await service.listRoles(req.auth.organizationId)})));router.post("/roles",validate(schemas.createRoleSchema),asyncHandler(async(req,res)=>res.status(201).json({success:true,data:await service.createRole(req.auth.organizationId,req.validated.body)})));router.patch("/roles/:roleId",validate(schemas.updateRoleSchema),asyncHandler(async(req,res)=>{await service.updateRole(req.auth.organizationId,req.validated.params.roleId,req.validated.body);res.status(204).end();}));router.get("/permissions",asyncHandler(async(_req,res)=>res.json({success:true,data:await service.listPermissions()})));return router;};
module.exports={createAccessRouter};