# rigweda/backend/src/modules/attendance/attendance.routes.js
const express=require("express");
const {asyncHandler}=require("../../middleware/async-handler");
const {validate}=require("../../middleware/validate");
const schemas=require("./attendance.validation");
const createAttendanceRouter=(service,permissions)=>{const router=express.Router();
  router.get("/metadata",asyncHandler(async(req,res)=>res.json({success:true,data:await service.metadata(req.auth.organizationId),requestId:req.id})));
  router.get("/me/today",asyncHandler(async(req,res)=>res.json({success:true,data:await service.myToday(req.auth.organizationId,req.auth.userId),requestId:req.id})));
  router.post("/me/punch",permissions.self,validate(schemas.punchSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.punch(req.auth.organizationId,req.auth.userId,req.validated.body,{ip:req.ip}),requestId:req.id})));
  router.get("/me/regularizations",validate(schemas.listRegularizationsSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.regularizations(req.auth.organizationId,req.validated.query,req.auth.userId),requestId:req.id})));
  router.post("/me/regularizations",permissions.self,validate(schemas.regularizationSchema),asyncHandler(async(req,res)=>res.status(201).json({success:true,data:await service.createRegularization(req.auth.organizationId,req.auth.userId,req.validated.body),requestId:req.id})));
  router.post("/me/regularizations/:requestId/cancel",permissions.self,validate(schemas.cancelSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.cancel(req.auth.organizationId,req.auth.userId,req.validated.params.requestId,req.validated.body.version),requestId:req.id})));
  router.get("/daily/summary",validate(schemas.dailySchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.dailySummary(req.auth.organizationId,req.validated.query.date),requestId:req.id})));
  router.get("/daily",validate(schemas.dailySchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.daily(req.auth.organizationId,req.validated.query),requestId:req.id})));
  router.get("/records/export",permissions.export,validate(schemas.rangeSchema),asyncHandler(async(req,res)=>{const csv=await service.export(req.auth.organizationId,req.validated.query);res.set({"Content-Type":"text/csv; charset=utf-8","Content-Disposition":`attachment; filename="attendance-${req.validated.query.dateFrom}-${req.validated.query.dateTo}.csv"`});res.send(`\uFEFF${csv}`);}));
  router.get("/records",validate(schemas.rangeSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.records(req.auth.organizationId,req.validated.query),requestId:req.id})));
  router.put("/records",permissions.manage,validate(schemas.manualSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.manual(req.auth.organizationId,req.validated.body,req.auth.userId),requestId:req.id})));
  router.get("/records/:recordId/history",permissions.manage,validate(schemas.recordIdSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.history(req.auth.organizationId,req.validated.params.recordId),requestId:req.id})));
  router.get("/records/:recordId",validate(schemas.recordIdSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.record(req.auth.organizationId,req.validated.params.recordId),requestId:req.id})));
  router.get("/policy",asyncHandler(async(req,res)=>res.json({success:true,data:await service.policy(req.auth.organizationId),requestId:req.id})));
  router.patch("/policy",permissions.manage,validate(schemas.policySchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.updatePolicy(req.auth.organizationId,req.validated.body,req.auth.userId),requestId:req.id})));
  router.get("/shifts",validate(schemas.listShiftsSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.shifts(req.auth.organizationId,req.validated.query),requestId:req.id})));
  router.post("/shifts",permissions.manage,validate(schemas.createShiftSchema),asyncHandler(async(req,res)=>res.status(201).json({success:true,data:await service.createShift(req.auth.organizationId,req.validated.body,req.auth.userId),requestId:req.id})));
  router.get("/shifts/:shiftId",validate(schemas.shiftIdSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.shift(req.auth.organizationId,req.validated.params.shiftId),requestId:req.id})));
  router.patch("/shifts/:shiftId",permissions.manage,validate(schemas.updateShiftSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.updateShift(req.auth.organizationId,req.validated.params.shiftId,req.validated.body,req.auth.userId),requestId:req.id})));
  router.get("/assignments",validate(schemas.listAssignmentsSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.assignments(req.auth.organizationId,req.validated.query),requestId:req.id})));
  router.post("/assignments",permissions.manage,validate(schemas.createAssignmentSchema),asyncHandler(async(req,res)=>res.status(201).json({success:true,data:await service.createAssignment(req.auth.organizationId,req.validated.body,req.auth.userId),requestId:req.id})));
  router.patch("/assignments/:assignmentId",permissions.manage,validate(schemas.updateAssignmentSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.updateAssignment(req.auth.organizationId,req.validated.params.assignmentId,req.validated.body),requestId:req.id})));
  router.get("/regularizations",permissions.approve,validate(schemas.listRegularizationsSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.regularizations(req.auth.organizationId,req.validated.query),requestId:req.id})));
  router.post("/regularizations/:requestId/review",permissions.approve,validate(schemas.reviewSchema),asyncHandler(async(req,res)=>res.json({success:true,data:await service.review(req.auth.organizationId,req.validated.params.requestId,req.validated.body,req.auth.userId),requestId:req.id})));
  return router;};
module.exports={createAttendanceRouter};